passport_bot/requirements.txt
slack-bolt==1.18.0
slack-sdk==3.27.1
playwright==1.40.0
python-dotenv==1.0.0
supabase==2.3.0
transformers==4.36.0
torch==2.9.1
pillow==11.0.0
asyncio==3.4.3